# hospital-service/docs/swagger.yaml
basePath: /
definitions:
...
  doctor.Doctor:
    properties:
      created_at:
        type: string
      doctor_id:
        type: string
      email:
        type: string
      full_name:
        type: string
      hospital_id:
        type: string
      image:
        type: string
      phone:
        type: string
      slug:
        type: string
      specialty:
        $ref: '#/definitions/enums.Specialty'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  enums.Specialty:
    enum:
    - ophthalmology
    - retina
    - cornea
    - glaucoma
    - refractive_surgery
    - pediatric_ophthalm
    - schedule_manager
    type: string
    x-enum-comments:
      SpecialtyCornea: Chuyên khoa giác mạc
      SpecialtyGlaucoma: Cườm nước
      SpecialtyOphthalmology: Nhãn khoa tổng quát
      SpecialtyPediatricOphthalm: Nhãn khoa nhi
      SpecialtyRefractiveSurgery: Phẫu thuật khúc xạ (LASIK)
      SpecialtyRetina: Chuyên khoa võng mạc
      SpecialtyScheduleManager: Quản lý lịch bác sĩ
    x-enum-descriptions:
    - Nhãn khoa tổng quát
    - Chuyên khoa võng mạc
    - Chuyên khoa giác mạc
    - Cườm nước
    - Phẫu thuật khúc xạ (LASIK)
    - Nhãn khoa nhi
    - Quản lý lịch bác sĩ
    x-enum-varnames:
    - SpecialtyOphthalmology
    - SpecialtyRetina
    - SpecialtyCornea
    - SpecialtyGlaucoma
    - SpecialtyRefractiveSurgery
    - SpecialtyPediatricOphthalm
    - SpecialtyScheduleManager
...
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8084
info:
  contact: {}
  title: Auth Service API
  version: "1.0"
paths:
...
  /doctors:
    get:
      description: Retrieve all doctors with optional filters
      parameters:
      - description: Filter by doctor name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by specialty (exact match)
        in: query
        name: specialty
        type: string
      - description: Filter by hospital ID (exact match)
        in: query
        name: hospital_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/doctor.Doctor'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all doctors
      tags:
      - Doctors
    post:
      consumes:
      - multipart/form-data
      description: Add a doctor with required info and optional avatar upload
      parameters:
      - description: User ID
        in: formData
        name: user_id
        required: true
        type: string
      - description: Full Name
        in: formData
        name: full_name
        required: true
        type: string
      - description: Specialty
        in: formData
        name: specialty
        required: true
        type: string
      - description: Hospital ID
        in: formData
        name: hospital_id
        required: true
        type: string
      - description: Phone
        in: formData
        name: phone
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: Avatar File
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Create a new doctor
      tags:
      - Doctors
  /doctors/{doctor_id}:
    delete:
      description: Delete doctor by doctor ID
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete doctor
      tags:
      - Doctors
    get:
      description: Retrieve doctor record using doctor ID
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doctor.Doctor'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get doctor by doctor ID
      tags:
      - Doctors
    put:
      consumes:
      - multipart/form-data
      description: Update doctor info and avatar
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: string
      - description: Full Name
        in: formData
        name: full_name
        type: string
      - description: Specialty
        in: formData
        name: specialty
        type: string
      - description: Hospital ID
        in: formData
        name: hospital_id
        type: string
      - description: Phone
        in: formData
        name: phone
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: Avatar File
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Update a doctor
      tags:
      - Doctors
...
  /doctors/hospital/{hospital_id}:
    get:
      description: Retrieve doctors for a specific hospital
      parameters:
      - description: Hospital ID
        in: path
        name: hospital_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/doctor.Doctor'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List doctors by hospital ID
      tags:
      - Doctors
  /doctors/slug/{slug}:
    get:
      description: Retrieve doctor record using slug
      parameters:
      - description: Doctor Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doctor.Doctor'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get doctor by slug
      tags:
      - Doctors
  /doctors/user/{user_id}:
    get:
      description: Retrieve doctor record using user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doctor.Doctor'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get doctor by user ID
      tags:
      - Doctors
...
swagger: "2.0"